Поправила функции на стрелочные в util.js, убрала одинаковую лишнюю функцию закрытия сообщения, удалила переменные DOM из импорта/экспорта
